feat(gateway): ship the Postgres migrations in the hosted container image
The hosted Gateway container publishes CcDirector.Gateway, but the Postgres EF
migrations live in a separate assembly (CcDirector.Gateway.Migrations.Postgres)
that references CcDirector.Gateway - so CcDirector.Gateway cannot reference it back
without a build cycle, and a container boot against Postgres would find no
migration set.

Add a thin host project, CcDirector.Gateway.Host, that references both
CcDirector.Gateway and the migrations assembly and is the container entry point.
The migrations DLL now ships in the image and Database.Migrate() resolves it by
name at runtime. This is the standard EF pattern for a separate migrations
assembly: the migrations assembly references the context assembly, and a startup
host references both.

To guarantee the container host and the local console host never drift, the whole
Gateway startup - the port parsing, the hosted platform-port resolution, and the
worker wiring - is extracted into a shared GatewayEntryPoint.Run(args) that both
executables call; each Program.cs is now a one-line delegation. No behavior change
to the local SQLite path.

The Dockerfile changes only the publish target and the entry point; EXPOSE, USER,
HOME, and CC_DIRECTOR_ROOT are unchanged.

Proven by a run: the host publish output carries the migrations DLL and references
it in the host deps.json; a boot smoke test resolves the migration set by name with
no database (CI-safe) and, when a real Postgres is configured, boots the real host
path and applies the migration on Supabase (skips cleanly when unset). The SQLite
path and its migration set are untouched.
